requirements.txt
fastapi
uvicorn
pdfminer.six
pandas
python-multipart
requests